rename @/data to @/crawler/data
